Bug #40559   assertion failed in check_binlog_magic
The reason of the bug is in that the test makes a trick with relay log files and
did not reset fully at the end.
If mtr does not restart the test the new SQL thread tried to work with the old time
session data.

Fixed with deploying RESET slave at the clean-up.