Bug#49477: Assertion `0' failed in ha_partition.cc:5530 with temporary table and partitions
It was possible to create temporary partitioned tables
via create table ... like ... (which is not allowed with
create temporary table). This lead to a new HA_EXTRA flag
(HA_EXTRA_MMAP) was sent to the partitioning handler,
which was caught on an assert in debug builds.

Solution was to check for partitioned tables when
doing create table ... like ... and disallow it.