BUG#50473: rpl_sync fails on windows debug enabled binaries
The test case was failing because it contained instructions
to close/reopen files, when they were in use. This raises
problems in windows. Example of such instruction:

---exec echo "failure" > $MYSQLD_SLAVE_DATADIR/$file

The test also contains commands that are not platform 
agnostic. Example:

--exec cat $MYSQLD_SLAVE_DATADIR/master.backup > \
           $MYSQLD_SLAVE_DATADIR/master.info

We fix this by just truncating the necessary file and write
"failure" into it (ie, without closing the file). The
platform specific instruction is removed from the test
case as it seems redundant.